requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Google Gemini
google-generativeai>=0.8.0

# Document Parsing
PyMuPDF==1.25.3
Pillow==11.1.0

# Vector Store (RAG)
chromadb==0.6.3

# Frontend
streamlit==1.41.1
plotly==6.5.2
requests>=2.31.0

# Testing
pytest==8.3.4
httpx==0.28.1
